Move AdvancedFraudSignalsTestRule to network-testing
